+++ agentlib/core/agent.py
@@ -17,6 +17,7 @@ from agentlib.core import (
     Environment,
     LocalDataBroker,
     RTDataBroker,
+    DirectCallbackDataBroker,
     BaseModule,
     DataBroker,
 )
@@ -61,6 +62,10 @@ class AgentConfig(BaseModel):
         ge=-1,
         description="Maximal number of waiting items in data-broker queues. "
         "Set to -1 for infinity",
+    )
+    use_direct_callback_databroker: bool = Field(
+        default=False,
+        description="If True, the `DirectCallbackDataBroker` will be used "
     )
 
     @field_validator("modules")
@@ -125,10 +130,16 @@ class Agent:
             env=self.env, name=f"{config.id}/DataBroker"
         )
         if env.config.rt:
+            if config.use_direct_callback_databroker:
+                raise ValueError("Can not use the direct callback databroker in real-time")
             self._data_broker = RTDataBroker(
                 env=env, logger=data_broker_logger, max_queue_size=config.max_queue_size
             )
             self.register_thread(thread=self._data_broker.thread)
+        elif config.use_direct_callback_databroker:
+            self._data_broker = DirectCallbackDataBroker(
+                logger=data_broker_logger
+            )
         else:
             self._data_broker = LocalDataBroker(
                 env=env, logger=data_broker_logger, max_queue_size=config.max_queue_size