+++ tests/test_agent.py
@@ -6,6 +6,7 @@ import os
 import threading
 import time
 import sys
+from copy import deepcopy
 
 from pydantic import ValidationError
 
@@ -82,6 +83,23 @@ class TestAgent(unittest.TestCase):
         with open(self.filepath, "w+") as file:
             json.dump(self._generate_ag_config()["modules"][0], file)
         AgentConfig(id=self.ag_id, modules=[self.filepath])
+
+    def test_modules_as_dict_or_list(self):
+        """Tests that agent configs support both list of modules and dict of modules."""
+        agent_config = self._generate_ag_config()
+        module_id = "test_id"
+        agent_config_with_dict = deepcopy(agent_config)
+        agent_config_with_dict["modules"] = {
+            module_id: agent_config_with_dict["modules"][0]
+        }
+        agent_config["modules"][0]["module_id"] = module_id
+        agent_list = AgentConfig(**agent_config)
+        agent_dict = AgentConfig(**agent_config_with_dict)
+        self.assertEqual(agent_list, agent_dict)
+
+        agent_config_with_dict["modules"]["test_id"]["module_id"] = "wrong_id"
+        with self.assertRaises(ConfigurationError):
+            _ = AgentConfig(**agent_config_with_dict)
 
     def test_getters(self):
         ag = Agent(config=self.ag_config, env=self.env)