+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## 0.8.9
+- Improve error message to include agent and module id in validation errors
+- Allow agent configs to specify modules as a dict instead of list, with dict keys corresponding to module_id. Old list still supported as normal.
 - Add DirectCallbackDataBroker option for easier debugging and learning how AgentLib Callbacks work #67
 
 ## 0.8.8